requirements.txt
quart
sqlmodel
asyncpg
sqlalchemy[asyncio]
python-dotenv
alembic
psycopg2-binary
